Typing your functional components
Objective: Component design can take some getting used to. In your applications and similar other applications, it is important to identify which part of your application can be broken down into smaller components. And how that smaller component can be reused. To reuse a component effectively, we need to make it dynamic by passing some props to that container.